[Gold V] Title: 입국심사, Time: 272 ms, Memory: 23452 KB -BaekjoonHub

+++ 백준/Gold/3079. 입국심사/입국심사.java
@@ -16,7 +16,7 @@ class Main{
 		while(left <= right){
 			long mid = (left + right) / 2;
 			long sum = 0;
-			for(long num : arr){
+			for(int num : arr){
 				sum += (mid/num);
 				if(sum >= m) 
 					break;
